libraries: Remove IVAS and VAI
VAI is now available on github and the yocto recipe will pull it from there.
IVAS will still need to be bundled as part of the design but it's better to
pull it into the bsp and change the SRC_URI to build from a tar ball instead
of git. This will be added back in later.