Jenkinsfile: Remove when condition from "Fix changelog"
The when condition in the "Fix changelog" stage prevents force-pushed pull
requests from building again. Remove it. As a side effect, the pull request
branch will also rebuild when triggered by branch indexing for example if
another pull request was merged and the base branch was updated as a result of
that.